+++ pkg/sync_compactor/compactor.go
@@ -5,14 +5,11 @@ import (
 	"errors"
 	"os"
 	"path"
-	"reflect"
 	"time"
 
 	reader_v2 "github.com/conductorone/baton-sdk/pb/c1/reader/v2"
 	"github.com/conductorone/baton-sdk/pkg/dotc1z"
 	c1zmanager "github.com/conductorone/baton-sdk/pkg/dotc1z/manager"
-	"google.golang.org/protobuf/proto"
-	"google.golang.org/protobuf/types/known/anypb"
 )
 
 type Compactor struct {
@@ -102,15 +99,15 @@ func (c *Compactor) doOneCompaction(ctx context.Context, tempDir string, base *C
 	}
 
 	// TODO: Use the runner when implementing the compaction logic
-	_ = &naiveCompactor{
+	runner := &naiveCompactor{
 		base:    baseFile,
 		applied: appliedFile,
 		dest:    newFile,
 	}
 
-	//if err := runner.processResourceTypes(ctx); err != nil {
-	//	return nil, err
-	//}
+	if err := runner.processResourceTypes(ctx); err != nil {
+		return nil, err
+	}
 
 	//
 	return nil, errors.New("NOT IMPLEMENTED")
@@ -120,88 +117,6 @@ type naiveCompactor struct {
 	base    *dotc1z.C1File
 	applied *dotc1z.C1File
 	dest    *dotc1z.C1File
-}
-
-type listRequest interface {
-	proto.Message
-	GetPageSize() uint32
-	GetPageToken() string
-	GetAnnotations() []*anypb.Any
-}
-
-type listResponse[T proto.Message] interface {
-	GetNextPageToken() string
-	GetAnnotations() []*anypb.Any
-	GetList() []T
-}
-
-// createRequest creates a new request object of type REQ using reflection
-func createRequest[REQ listRequest]() REQ {
-	reqType := reflect.TypeOf((*REQ)(nil)).Elem()
-	reqPtrValue := reflect.New(reqType)
-	return reqPtrValue.Interface().(REQ)
-}
-
-// setFieldIfValid sets a field in a struct if it exists and can be set
-func setFieldIfValid(obj interface{}, fieldName string, setValue func(reflect.Value)) {
-	val := reflect.ValueOf(obj)
-	if val.Kind() != reflect.Ptr || val.IsNil() {
-		return
-	}
-
-	field := val.Elem().FieldByName(fieldName)
-	if field.IsValid() && field.CanSet() {
-		setValue(field)
-	}
-}
-
-// setPageSize sets the PageSize field in a request to the specified value
-func setPageSize(req listRequest, size uint64) {
-	setFieldIfValid(req, "PageSize", func(field reflect.Value) {
-		field.SetUint(size)
-	})
-}
-
-// setPageToken sets the PageToken field in a request to the specified token
-func setPageToken(req listRequest, token string) {
-	setFieldIfValid(req, "PageToken", func(field reflect.Value) {
-		field.SetString(token)
-	})
-}
-
-func UnrollAll[T proto.Message, REQ listRequest, RESP listResponse[T]](ctx context.Context, list func(context.Context, REQ) (RESP, error)) ([]T, error) {
-	var allResults []T
-
-	// Create a new request using reflection
-	req := createRequest[REQ]()
-
-	// Set initial page size
-	setPageSize(req, 100) // Set a reasonable default page size
-
-	var nextPageToken string
-	for {
-		// Set the page token for the current request if needed
-		if nextPageToken != "" {
-			setPageToken(req, nextPageToken)
-		}
-
-		// Call the list function with the current request
-		resp, err := list(ctx, req)
-		if err != nil {
-			return nil, err
-		}
-
-		// Collect the results
-		allResults = append(allResults, resp.GetList()...)
-
-		// Check if there are more pages
-		nextPageToken = resp.GetNextPageToken()
-		if nextPageToken == "" {
-			break // No more pages
-		}
-	}
-
-	return allResults, nil
 }
 
 func (c *naiveCompactor) processResourceTypes(ctx context.Context) error {
